venueControllr

+++ src/main/java/org/example/_10_pon_cahannarith_pp_spring_homework003/model/entity/Venue.java
@@ -6,7 +6,7 @@ import lombok.*;
 @NoArgsConstructor
 @AllArgsConstructor
 public class Venue {
-    private Long venueId;
+    private Integer venueId;
     private String venueName;
     private String location;
